[Style/#48]: 팔로우 fallback 적용
