+++ src/constants/images.ts
@@ -27,6 +27,7 @@ export const IMAGES = {
     },
     memo_frog: `${BASE_PATH}/frog/memo-frog.svg`,
     more_feed: `${BASE_PATH}/frog/morefeed-frog.svg`,
+    empty_dot: `${BASE_PATH}/frog/frog-empty-dot.svg`,
   },
   test: {
     result: {
